refactor(render): modularize core/render directory structure
Move codebase files into backend/, graph/, and resource/ subfolders and update include directives.
