+++ core/Application.h
@@ -5,13 +5,13 @@
 #include "AssetManager.h"
 #include "Layer.h"
 #include "Window.h"
-#include "render/BindGroupManager.h"
-#include "render/MaterialManager.h"
-#include "render/MeshManager.h"
-#include "render/PipelineManager.h"
+#include "render/backend/BindGroupManager.h"
+#include "render/resource/MaterialManager.h"
+#include "render/resource/MeshManager.h"
+#include "render/backend/PipelineManager.h"
 #include "render/SceneRenderer.h"
-#include "render/ShaderManager.h"
-#include "render/TextureManager.h"
+#include "render/resource/ShaderManager.h"
+#include "render/resource/TextureManager.h"
 #include "render/render.h"
 
 namespace core {
