+++ core/CMakeLists.txt
@@ -13,19 +13,19 @@ set(SOURCES
     "AssetManager.h"
     "AssetManager.cpp" 
     "render/util.h"
-    "render/GpuResource.h"
-    "render/Mesh.h" 
-    "render/Mesh.cpp" 
+    "render/resource/GpuResource.h"
+    "render/resource/Mesh.h" 
+    "render/resource/Mesh.cpp" 
     "ResourcePool.h" 
-    "render/RenderGraph.h"
-    "render/RenderGraph.cpp" 
-    "render/LayoutCache.h"
-    "render/LayoutCache.cpp"
-    "render/PipelineManager.h"
-    "render/PipelineManager.cpp"
-    "render/ShaderAsset.h" "render/ShaderAsset.cpp"
-    "render/ShaderManager.h" "render/ShaderManager.cpp"
-    "render/Material.h" "render/Material.cpp"
+    "render/graph/RenderGraph.h"
+    "render/graph/RenderGraph.cpp" 
+    "render/backend/LayoutCache.h"
+    "render/backend/LayoutCache.cpp"
+    "render/backend/PipelineManager.h"
+    "render/backend/PipelineManager.cpp"
+    "render/resource/ShaderAsset.h" "render/resource/ShaderAsset.cpp"
+    "render/resource/ShaderManager.h" "render/resource/ShaderManager.cpp"
+    "render/resource/Material.h" "render/resource/Material.cpp"
     "memory/StridedSpan.h"
     "memory/StridedSpan.cpp" 
     "util/Load.h" 
@@ -39,19 +39,19 @@ set(SOURCES
     "import/Importer.h"
     "import/GLTFImporter.h" "import/GLTFImporter.cpp"
     "import/ShdrImporter.h" "import/ShdrImporter.cpp"
-    "render/MaterialManager.h" "render/MaterialManager.cpp"
-    "render/Texture.h" 
-    "render/TextureManager.h" "render/TextureManager.cpp"
-    "render/MeshManager.h"  "render/MeshManager.cpp"
-    "render/Model.h"
-    "render/VertexLayoutManager.h" "render/VertexLayoutManager.cpp" 
+    "render/resource/MaterialManager.h" "render/resource/MaterialManager.cpp"
+    "render/resource/Texture.h" 
+    "render/resource/TextureManager.h" "render/resource/TextureManager.cpp"
+    "render/resource/MeshManager.h"  "render/resource/MeshManager.cpp"
+    "render/resource/Model.h"
+    "render/resource/VertexLayoutManager.h" "render/resource/VertexLayoutManager.cpp" 
     "wgx/types.h" 
     "render/SceneRenderer.h" "render/SceneRenderer.cpp"
     "Scene.h"
-    "render/IRenderPass.cpp" 
+    "render/graph/IRenderPass.cpp" 
     "render/pass/DeferredGBufferPass.h" "render/pass/DeferredGBufferPass.cpp"
     "render/pass/DeferredLightingPass.h" "render/pass/DeferredLightingPass.cpp"
-    "render/pass/ForwardRenderPass.h" "render/pass/ForwardRenderPass.cpp" "render/BindGroupManager.h" "render/BindGroupManager.cpp" "render/BindGroupFactory.h" )
+    "render/pass/ForwardRenderPass.h" "render/pass/ForwardRenderPass.cpp" "render/backend/BindGroupManager.h" "render/backend/BindGroupManager.cpp" "render/backend/BindGroupFactory.h" )
 
 include (../cmake/ShaderCompiler.cmake)
