+++ core/import/ShdrImporter.h
@@ -4,7 +4,7 @@
 #include <Common.h>
 #include "Importer.h"
 #include "ShaderAssetFormat.h"
-#include "render/ShaderAsset.h"
+#include "render/resource/ShaderAsset.h"
 namespace core::render {
 class ShaderReflection;
 }
