+++ core/render/resource/GpuResource.h
