+++ core/render/resource/Material.cpp
@@ -1,4 +1,4 @@
-#include "IRenderPass.h"
+#include "render/graph/IRenderPass.h"
 #include "Material.h"
 #include "render/util.h"
