+++ core/render/resource/MaterialManager.cpp
