+++ core/render/resource/ShaderManager.cpp
@@ -1,7 +1,7 @@
 #include "ShaderManager.h"
 #include <ranges>
 #include "asset/StandardPBR.h"
-#include "util.h"
+#include "render/util.h"
 
 namespace core::render {
